feat: :sparkles: Status filter
The database in which the tasks will be collected was linked to the project, and two functions were added. These functions are responsible for printing the data in the HTML file: one prints tasks with a default style if their status is 'ready', while the other prints tasks with a style that strikes through the task name and makes the color more transparent if their status is 'onHold'. This distinction aims to visually differentiate between tasks marked as 'ready' and those on hold